requirements.txt
numpy~=1.21.4
Pillow==8.4.0
matplotlib==3.3.4
torch==1.10.0
torchvision==0.11.1
cleverhans==4.0.0
facenet-pytorch==2.5.2
kiwisolver==1.3.1
cycler==0.11.0
python-dateutil==2.8.2
pyparsing==3.0.4
typing-extensions==3.10.0.2
dataclasses==0.6
tensorflow-probability==0.14.1
scipy==1.5.4
nose==1.3.7
easydict==1.9
six==1.16.0
pycodestyle==2.8.0
joblib==1.1.0
mnist==0.2.2
absl-py==0.15.0
requests==2.26.0
gast==0.5.2
dm-tree==0.1.6
decorator==5.1.0
cloudpickle==2.0.0
urllib3==1.26.7
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
torchattacks==3.2.2
rich~=10.15.1
ffmpeg-python
facelib~=1.2.1
opencv-python~=4.5.4.60
scikit-image~=0.19.0